Changes from qmk pull request
outdated file removed
oled in keyboard level
info.json removed old stuff
