+++ keyboards/rarepotato8de/3x3macropad/info.json
@@ -2,11 +2,10 @@
     "manufacturer": "RarePotato8DE",
     "keyboard_name": "3x3macropad",
     "maintainer": "RarePotato8DE",
-    "bootloader": "caterina",
+    "development_board": "promicro",
     "build": {
         "lto": true
     },
-    "debounce": 5,
     "diode_direction": "COL2ROW",
     "encoder": {
         "rotary": [
@@ -15,20 +14,16 @@
     },
     "features": {
         "bootmagic": true,
-        "combo": false,
         "encoder": true,
         "extrakey": true,
-        "key_override": false,
         "nkro": true,
         "oled": true,
-        "rgblight": true,
-        "tap_dance": false
+        "rgblight": true
     },
     "matrix_pins": {
         "cols": ["D7", "B3", "B2"],
         "rows": ["F7", "B5", "B6"]
     },
-    "processor": "atmega32u4",
     "rgblight": {
         "animations": {
             "alternating": true,
@@ -38,8 +33,7 @@
             "snake": true,
             "static_gradient": true
         },
-        "led_count": 6,
-        "pin": "E6"
+        "led_count": 6
     },
     "url": "https://github.com/rarepotato8de/3x3macropad",
     "usb": {
@@ -52,17 +46,16 @@
     },
     "layouts": {
         "LAYOUT": {
-            "c_macro": true,
             "layout": [
-                {"label": "k00", "matrix": [0, 0], "x": 0, "y": 0, "w": 1},
-                {"label": "k01", "matrix": [0, 1], "x": 1, "y": 0, "w": 1},
-                {"label": "k02", "matrix": [0, 2], "x": 2, "y": 0, "w": 1},
-                {"label": "k10", "matrix": [1, 0], "x": 0, "y": 1, "w": 1},
-                {"label": "k11", "matrix": [1, 1], "x": 1, "y": 1, "w": 1},
-                {"label": "k12", "matrix": [1, 2], "x": 2, "y": 1, "w": 1},
-                {"label": "k20", "matrix": [2, 0], "x": 0, "y": 2, "w": 1},
-                {"label": "k21", "matrix": [2, 1], "x": 1, "y": 2, "w": 1},
-                {"label": "k22", "matrix": [2, 2], "x": 2, "y": 2, "w": 1}
+                {"label": "k00", "matrix": [0, 0], "x": 0, "y": 0},
+                {"label": "k01", "matrix": [0, 1], "x": 1, "y": 0},
+                {"label": "k02", "matrix": [0, 2], "x": 2, "y": 0},
+                {"label": "k10", "matrix": [1, 0], "x": 0, "y": 1},
+                {"label": "k11", "matrix": [1, 1], "x": 1, "y": 1},
+                {"label": "k12", "matrix": [1, 2], "x": 2, "y": 1},
+                {"label": "k20", "matrix": [2, 0], "x": 0, "y": 2},
+                {"label": "k21", "matrix": [2, 1], "x": 1, "y": 2},
+                {"label": "k22", "matrix": [2, 2], "x": 2, "y": 2}
             ]
         }
     }
